# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Trapa Center API",
    "description": "駐車場入出庫トラッキングシステム「トラパ」のセンターAPI。エッジデバイスからの入出庫イベント登録、ヘルスチェック、デバイスへのコマンド（再起動など）配信をポーリング方式で提供する。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "管理者ログイン",
        "description": "ユーザー名とパスワードを検証し、アクセストークンを返す。同時にリフレッシュ\nトークンをHTTP OnlyクッキーとしてセットするAdmin console向けのログイン。",
        "operationId": "login_api_v1_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "ログイン中のユーザー情報取得",
        "operationId": "me_api_v1_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
    "/api/v1/heartbeat": {
      "post": {
        "tags": [
          "device"
        ],
        "summary": "ハートビート送信",
        "description": "デバイスが定期的に呼び出す生存確認エンドポイント。ステータスを報告すると\n同時に、前回のポーリング以降にキューされたコマンド（再起動など）を受け取る。\nNAT配下などサーバーから直接到達できないデバイスに対して、サーバーが\nコマンドを届けられる唯一の経路。",
        "operationId": "heartbeat_api_v1_heartbeat_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HeartbeatOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/commands/{command_id}/ack": {
      "post": {
        "tags": [
          "device"
        ],
        "summary": "コマンド実行結果の報告",
        "description": "デバイスが受け取ったコマンド（再起動など）の実行結果（成功/失敗）を報告する。",
        "operationId": "ack_command_api_v1_commands__command_id__ack_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "command_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Command Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandAck"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/devices": {
      "get": {
        "tags": [
          "devices"
        ],
        "summary": "デバイス一覧取得",
        "description": "登録済みデバイスの一覧と、最終通信時刻から判定したオンライン状態を返す。",
        "operationId": "list_devices_api_v1_devices_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DeviceOut"
                  },
                  "type": "array",
                  "title": "Response List Devices Api V1 Devices Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "devices"
        ],
        "summary": "デバイス登録",
        "description": "新しいエッジデバイスを登録し、平文のAPIキーを一度だけ返す。サーバーには\nSHA-256ハッシュのみ保存されるため、後から再取得することはできない。",
        "operationId": "register_device_api_v1_devices_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceCreateOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/devices/{device_id}": {
      "get": {
        "tags": [
          "devices"
        ],
        "summary": "デバイス詳細取得",
        "description": "指定したデバイスの詳細とオンライン状態を返す。",
        "operationId": "get_device_api_v1_devices__device_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "devices"
        ],
        "summary": "デバイス情報の更新",
        "description": "デバイスコード・表示名・紐づく駐車場を更新する。指定したフィールドのみ更新される。\nAPIキー自体はここでは変更されない。",
        "operationId": "update_device_api_v1_devices__device_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "devices"
        ],
        "summary": "デバイス削除",
        "description": "デバイスを削除する。紐づく入出庫イベント・コマンド履歴も合わせて削除される。",
        "operationId": "delete_device_api_v1_devices__device_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/devices/{device_id}/commands": {
      "post": {
        "tags": [
          "devices"
        ],
        "summary": "コマンドをキューに登録",
        "description": "デバイスへのコマンド（再起動など）をキューに積む。デバイスは次回の\n/heartbeat 呼び出し時にこれを受け取る。サーバーからデバイスへの直接\nプッシュは行わない。",
        "operationId": "queue_command_api_v1_devices__device_id__commands_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CommandCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommandOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "devices"
        ],
        "summary": "コマンド履歴取得",
        "description": "指定したデバイスに対して発行したコマンドの履歴を新しい順に返す。",
        "operationId": "list_commands_api_v1_devices__device_id__commands_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CommandOut"
                  },
                  "title": "Response List Commands Api V1 Devices  Device Id  Commands Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CommandAck": {
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "completed",
              "failed"
            ],
            "title": "Status",
            "description": "コマンドの実行結果"
          },
          "result_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Message",
            "description": "実行結果の詳細メッセージ（任意）"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "CommandAck"
      },
      "CommandCreate": {
        "properties": {
          "command_type": {
            "type": "string",
            "const": "restart",
            "title": "Command Type",
            "description": "コマンド種別（現在は再起動のみ対応）"
          },
          "requested_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested By",
            "description": "コマンドを発行した主体（管理者名など、任意）"
          }
        },
        "type": "object",
        "required": [
          "command_type"
        ],
        "title": "CommandCreate"
      },
      "CommandOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "device_id": {
            "type": "integer",
            "title": "Device Id",
            "description": "コマンドの宛先デバイスID"
          },
          "command_type": {
            "type": "string",
            "title": "Command Type",
            "description": "コマンド種別"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "コマンドの状態（pending=未配信 / delivered=配信済み / completed=完了 / failed=失敗）"
          },
          "requested_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested By"
          },
          "result_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Message"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "コマンドがキューに登録された日時"
          },
          "delivered_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delivered At",
            "description": "デバイスがハートビートで受け取った日時"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At",
            "description": "デバイスが実行結果を報告した日時"
          }
        },
        "type": "object",
        "required": [
          "id",
          "device_id",
          "command_type",
          "status",
          "requested_by",
          "result_message",
          "created_at",
          "delivered_at",
          "completed_at"
        ],
        "title": "CommandOut"
      },
      "DeviceCreate": {
        "properties": {
          "device_code": {
            "type": "string",
            "title": "Device Code",
            "description": "デバイスの一意な識別コード（例: trapa-dev1）"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "表示用のデバイス名（任意）"
          },
          "parking_lot_id": {
            "type": "integer",
            "title": "Parking Lot Id",
            "description": "このデバイスが紐づく駐車場のID"
          }
        },
        "type": "object",
        "required": [
          "device_code",
          "parking_lot_id"
        ],
        "title": "DeviceCreate"
      },
      "DeviceCreateOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "device_code": {
            "type": "string",
            "title": "Device Code"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "parking_lot_id": {
            "type": "integer",
            "title": "Parking Lot Id"
          },
          "api_key": {
            "type": "string",
            "title": "Api Key",
            "description": "デバイス認証用APIキー（平文）。このレスポンスでのみ取得可能で、以降は再取得できない"
          }
        },
        "type": "object",
        "required": [
          "id",
          "device_code",
          "name",
          "parking_lot_id",
          "api_key"
        ],
        "title": "DeviceCreateOut"
      },
      "DeviceOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "device_code": {
            "type": "string",
            "title": "Device Code"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "parking_lot_id": {
            "type": "integer",
            "title": "Parking Lot Id"
          },
          "last_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Status",
            "description": "直近のハートビートで報告された状態"
          },
          "last_seen_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At",
            "description": "最終通信日時（ハートビートまたはイベント受信）"
          },
          "online": {
            "type": "boolean",
            "title": "Online",
            "description": "最終通信からの経過時間がしきい値以内であればtrue"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "device_code",
          "name",
          "parking_lot_id",
          "last_status",
          "last_seen_at",
          "online",
          "created_at"
        ],
        "title": "DeviceOut"
      },
      "DeviceUpdate": {
        "properties": {
          "device_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Code",
            "description": "デバイスの一意な識別コード（例: trapa-dev1）"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "表示用のデバイス名"
          },
          "parking_lot_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parking Lot Id",
            "description": "このデバイスが紐づく駐車場のID"
          }
        },
        "type": "object",
        "title": "DeviceUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "HeartbeatIn": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "デバイスが自己申告する状態（例: ok, degraded）",
            "default": "ok"
          }
        },
        "type": "object",
        "title": "HeartbeatIn"
      },
      "HeartbeatOut": {
        "properties": {
          "server_time": {
            "type": "string",
            "format": "date-time",
            "title": "Server Time",
            "description": "サーバー側の受信日時"
          },
          "commands": {
            "items": {
              "$ref": "#/components/schemas/CommandOut"
            },
            "type": "array",
            "title": "Commands",
            "description": "このハートビートで配信された、実行待ちのコマンド一覧"
          }
        },
        "type": "object",
        "required": [
          "server_time",
          "commands"
        ],
        "title": "HeartbeatOut"
      },
      "LoginIn": {
        "properties": {
          "username": {
            "type": "string",
            "title": "Username",
            "description": "管理者ユーザー名"
          },
          "password": {
            "type": "string",
            "title": "Password",
            "description": "パスワード"
          }
        },
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "title": "LoginIn"
      },
...
      "TokenOut": {
        "properties": {
          "access_token": {
            "type": "string",
            "title": "Access Token",
            "description": "短命なJWTアクセストークン。フロントエンドはメモリ上にのみ保持する"
          },
          "token_type": {
            "type": "string",
            "title": "Token Type",
            "default": "bearer"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In",
            "description": "アクセストークンの有効期限（秒）"
          }
        },
        "type": "object",
        "required": [
          "access_token",
          "expires_in"
        ],
        "title": "TokenOut"
      },
      "UserOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          }
        },
        "type": "object",
        "required": [
          "id",
          "username"
        ],
        "title": "UserOut"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "description": "Admin console access token (from /auth/login)",
        "scheme": "bearer"
      },
      "APIKeyHeader": {
        "type": "apiKey",
        "description": "Edge device API key",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
}